Fix random crash when logging in and added ticket listing

+++ app/src/main/java/com/example/unieventos/ui/screens/SignUpScreen.kt
@@ -175,7 +175,7 @@ fun SignUpForm(
             onClick = {
                 usersViewModel.createUser(
                     User(
-                        cedula.toInt(),
+                        cedula,
                         cedula,
                         nombre,
                         direccion,
